Sat, Dec 16, 2023, 6:38 PM +00:00

+++ credit/credit.c
@@ -11,7 +11,7 @@ int main(void)
     //Visa = 4xxxxxxxxxxxx(xxx)
 
     int digit[16];
-    double divisor = 10;
+    long divisor = 10;
     int i;
 
     for (i = 1; i <= 16; i++)
